finish

+++ app/controllers/books_controller.rb
@@ -23,11 +23,10 @@ class BooksController < ApplicationController
       # '/books/new'
     else
       @books = Book.all
-      render :new
+      flash.now[:notice] ="error"
+      render :index
     end
   end
-
-  
 
   def show
     @book = Book.find(params[:id])
@@ -38,15 +37,26 @@ class BooksController < ApplicationController
   end
   
   def update
-    book = Book.find(params[:id])
-    book.update(book_params)
-    redirect_to book_path(book.id)
+    @book = Book.find(params[:id])
+    if @book.update(book_params)
+      flash[:notice] = "Book was successfully created"
+      redirect_to book_path(@book.id)
+    else
+      flash.now[:notice] ="error"
+      render :edit
+    end
   end
   
   def destroy
     book = Book.find(params[:id])  # データ（レコード）を1件取得
-    book.destroy  # データ（レコード）を削除
-    redirect_to '/books/new'  # 投稿一覧画面へリダイレクト
+    if
+      book.destroy  # データ（レコード）を削除
+      flash[:notice] = "Book was successfully created"
+      redirect_to '/books'  # 投稿一覧画面へリダイレクト
+    else
+      flash.now[:notice] ="error"
+      render :index
+    end
   end
   
   private
